+++ product_code/certificate_service/api/src/main/scala/de/upb/cs/uc4/certificate/model/PostMessageCSR.scala
@@ -7,18 +7,14 @@ import scala.concurrent.{ ExecutionContext, Future }
 
 case class PostMessageCSR(certificateSigningRequest: String, encryptedPrivateKey: EncryptedPrivateKey) {
 
-  def validate(implicit ec: ExecutionContext): Future[Seq[SimpleError]] = Future {
-    val allRegex = """[\s\S]{1,2000}""".r
-
+  def validate(implicit ec: ExecutionContext): Future[Seq[SimpleError]] = encryptedPrivateKey.validate.map { result =>
     var errors = List[SimpleError]()
 
-    encryptedPrivateKey match {
-      case EncryptedPrivateKey("", "", "") =>
-      case EncryptedPrivateKey(allRegex(_*), allRegex(_*), allRegex(_*)) =>
-      case _ =>
-        errors :+= SimpleError("encryptedPrivateKey", "Either all fields must be empty or no fields must be empty.")
+    if (certificateSigningRequest.trim.isEmpty) {
+      errors :+= SimpleError("certificateSigningRequest", "The certificateSigningRequest must be set.")
     }
-    errors
+
+    errors ++ result
   }
 }