+++ product_code/shared/client/src/main/scala/de/upb/cs/uc4/shared/client/exceptions/DetailedError.scala
@@ -3,7 +3,10 @@ package de.upb.cs.uc4.shared.client.exceptions
 import de.upb.cs.uc4.shared.client.exceptions.ErrorType.ErrorType
 import play.api.libs.json._
 
-case class DetailedError(`type`: ErrorType, title: String, invalidParams: Seq[SimpleError]) extends UC4Error
+case class DetailedError(`type`: ErrorType, title: String, invalidParams: Seq[SimpleError]) extends UC4Error {
+
+  override def toString: String = super.toString + "\n" + invalidParams.mkString("\t", "\n\t", "")
+}
 
 object DetailedError {
   implicit val format: Format[DetailedError] = Json.format