+++ product_code/shared/client/src/main/scala/de/upb/cs/uc4/shared/client/exceptions/SimpleError.scala
@@ -2,7 +2,10 @@ package de.upb.cs.uc4.shared.client.exceptions
 
 import play.api.libs.json.{ Format, Json }
 
-case class SimpleError(name: String, reason: String)
+case class SimpleError(name: String, reason: String) {
+
+  override def toString: String = name + " : " + reason
+}
 
 object SimpleError {
   implicit val format: Format[SimpleError] = Json.format