+++ product_code/user_service/impl/src/main/scala/de/upb/cs/uc4/user/impl/UserServiceImpl.scala
@@ -1,5 +1,7 @@
 package de.upb.cs.uc4.user.impl
 
+import java.util
+
 import akka.cluster.sharding.typed.scaladsl.{ ClusterSharding, EntityRef }
 import akka.util.{ ByteString, Timeout }
 import akka.{ Done, NotUsed }
@@ -13,6 +15,7 @@ import com.lightbend.lagom.scaladsl.server.ServerServiceCall
 import com.typesafe.config.Config
 import de.upb.cs.uc4.authentication.api.AuthenticationService
 import de.upb.cs.uc4.authentication.model.{ AuthenticationRole, JsonUsername }
+import de.upb.cs.uc4.image.api.ImageProcessingService
 import de.upb.cs.uc4.shared.client.exceptions._
 import de.upb.cs.uc4.shared.client.kafka.EncryptionContainer
 import de.upb.cs.uc4.shared.server.Hashing
@@ -38,13 +41,20 @@ import scala.concurrent.{ Await, ExecutionContext, Future, TimeoutException }
 class UserServiceImpl(
     clusterSharding: ClusterSharding, persistentEntityRegistry: PersistentEntityRegistry,
     readSide: ReadSide, processor: UserEventProcessor, database: UserDatabase,
-    authentication: AuthenticationService, kafkaEncryptionUtility: KafkaEncryptionUtility
+    authentication: AuthenticationService, kafkaEncryptionUtility: KafkaEncryptionUtility, imageProcessing: ImageProcessingService
 )(implicit ec: ExecutionContext, config: Config) extends UserService {
   readSide.register(processor)
 
   private final val log: Logger = LoggerFactory.getLogger("Shared")
 
   private lazy val defaultProfilePicture = ByteStreams.toByteArray(getClass.getResourceAsStream("/DefaultProfile.png"))
+  private lazy val defaultThumbnail = ByteStreams.toByteArray(getClass.getResourceAsStream("/DefaultThumbnail.png"))
+  private lazy val supportedTypes: util.List[String] = config.getStringList("uc4.image.supportedTypes")
+  private lazy val convertedTypes: util.List[String] = config.getStringList("uc4.image.convertedTypes")
+  private lazy val profileWidth: Int = config.getInt("uc4.image.profileWidth")
+  private lazy val profileHeight: Int = config.getInt("uc4.image.profileHeight")
+  private lazy val thumbnailWidth: Int = config.getInt("uc4.image.thumbnailWidth")
+  private lazy val thumbnailHeight: Int = config.getInt("uc4.image.thumbnailHeight")
 
   /** Looks up the entity for the given ID */
   private def entityRef(id: String): EntityRef[UserCommand] =
@@ -86,7 +96,6 @@ class UserServiceImpl(
           ResponseHeader(200, MessageProtocol.empty, List()),
           GetAllUsersResponse(students._2, lecturers._2, admins._2)
         )
-
       }
     }
 
@@ -132,7 +141,7 @@ class UserServiceImpl(
 
       // Check, if username errors exist, since entityRef might fail if username is incorrect
       if (validationErrors.map(_.name).contains(userVariableName + ".username")) {
-        throw new UC4Exception(422, DetailedError(ErrorType.Validation, validationErrors))
+        throw new UC4NonCriticalException(422, DetailedError(ErrorType.Validation, validationErrors))
       }
 
       val ref = entityRef(postMessageUser.getUser.username)
@@ -143,7 +152,7 @@ class UserServiceImpl(
         }
 
         if (validationErrors.nonEmpty) {
-          throw new UC4Exception(422, DetailedError(ErrorType.Validation, validationErrors))
+          throw new UC4NonCriticalException(422, DetailedError(ErrorType.Validation, validationErrors))
         }
 
         ref.ask[Confirmation](replyTo => CreateUser(postMessageUser.getUser, replyTo))
@@ -168,8 +177,8 @@ class UserServiceImpl(
                           throw deletionException
                       }
                 }
-            case RejectedWithError(code, errorResponse) =>
-              throw new UC4Exception(code, errorResponse)
+            case RejectedWithError(code, reason) =>
+              throw UC4Exception(code, reason)
           }
       }
     }
@@ -202,7 +211,7 @@ class UserServiceImpl(
             }
 
             if (validationErrors.map(_.name).contains("username")) {
-              throw new UC4Exception(422, DetailedError(ErrorType.Validation, validationErrors))
+              throw new UC4NonCriticalException(422, DetailedError(ErrorType.Validation, validationErrors))
             }
 
             val ref = entityRef(user.username)
@@ -211,7 +220,7 @@ class UserServiceImpl(
                 if (optUser.isEmpty) {
                   // Add to validation errors, and throw prematurely since uneditable fields are uncheckable
                   validationErrors :+= SimpleError("username", "Username not in use.")
-                  throw new UC4Exception(422, DetailedError(ErrorType.Validation, validationErrors))
+                  throw new UC4NonCriticalException(422, DetailedError(ErrorType.Validation, validationErrors))
                 }
                 val oldUser = optUser.get
                 // Other users than admins can only edit specified fields
@@ -220,16 +229,16 @@ class UserServiceImpl(
                   editErrors ++= oldUser.checkProtectedFields(user)
                 }
                 if (editErrors.nonEmpty) {
-                  throw new UC4Exception(422, DetailedError(ErrorType.UneditableFields, editErrors))
+                  throw new UC4NonCriticalException(422, DetailedError(ErrorType.UneditableFields, editErrors))
                 }
                 if (validationErrors.nonEmpty) {
-                  throw new UC4Exception(422, DetailedError(ErrorType.Validation, validationErrors))
+                  throw new UC4NonCriticalException(422, DetailedError(ErrorType.Validation, validationErrors))
                 }
 
                 oldUser match {
-                  case _: Student if !user.isInstanceOf[Student] => throw new UC4Exception(400, InformativeError(ErrorType.UnexpectedEntity, "Expected Student, but received non-Student"))
-                  case _: Lecturer if !user.isInstanceOf[Lecturer] => throw new UC4Exception(400, InformativeError(ErrorType.UnexpectedEntity, "Expected Lecturer, but received non-Lecturer"))
-                  case _: Admin if !user.isInstanceOf[Admin] => throw new UC4Exception(400, InformativeError(ErrorType.UnexpectedEntity, "Expected Admin, but received non-Admin"))
+                  case _: Student if !user.isInstanceOf[Student] => throw new UC4NonCriticalException(400, InformativeError(ErrorType.UnexpectedEntity, "Expected Student, but received non-Student"))
+                  case _: Lecturer if !user.isInstanceOf[Lecturer] => throw new UC4NonCriticalException(400, InformativeError(ErrorType.UnexpectedEntity, "Expected Lecturer, but received non-Lecturer"))
+                  case _: Admin if !user.isInstanceOf[Admin] => throw new UC4NonCriticalException(400, InformativeError(ErrorType.UnexpectedEntity, "Expected Admin, but received non-Admin"))
                   case _ =>
                 }
 
@@ -237,8 +246,8 @@ class UserServiceImpl(
                   .map {
                     case Accepted => // Update successful
                       (ResponseHeader(200, MessageProtocol.empty, List()), Done)
-                    case RejectedWithError(code, errorResponse) => //Update failed
-                      throw new UC4Exception(code, errorResponse)
+                    case RejectedWithError(code, reason) => //Update failed
+                      throw UC4Exception(code, reason)
                   }
             }
         }
@@ -254,8 +263,8 @@ class UserServiceImpl(
           .map {
             case Accepted => // Update Successful
               (ResponseHeader(200, MessageProtocol.empty, List()), Done)
-            case Rejected("A user with the given username does not exist.") => // Already exists
-              throw UC4Exception.NotFound
+            case RejectedWithError(code, reason) =>
+              throw UC4Exception(code, reason)
           }
     }
   }
@@ -335,7 +344,7 @@ class UserServiceImpl(
       val ref = entityRef(matriculationUpdate.username)
       ref.ask[Confirmation](replyTo => UpdateLatestMatriculation(matriculationUpdate.semester, replyTo)).map {
         case Accepted => Done
-        case RejectedWithError(error, reason) => throw new UC4Exception(error, reason)
+        case RejectedWithError(error, reason) => throw UC4Exception(error, reason)
       }
   }
 
@@ -392,6 +401,22 @@ class UserServiceImpl(
     }
   }
 
+  /** Gets the thumbnail of the user */
+  override def getThumbnail(username: String): ServiceCall[NotUsed, ByteString] = authenticated[NotUsed, ByteString](AuthenticationRole.All: _*) {
+    ServerServiceCall {
+      (header, _) =>
+        database.getThumbnail(username).flatMap {
+          case Some((array, contentType)) =>
+            Future.successful(ResponseHeader(200, MessageProtocol(contentType = Some(s"$contentType; charset=UTF-8")), List()), ByteString(array))
+          case None =>
+            getUser(username).invokeWithHeaders(header, NotUsed).map {
+              _ =>
+                (ResponseHeader(200, MessageProtocol(contentType = Some(s"image/png; charset=UTF-8")), List()), ByteString(defaultThumbnail))
+            }
+        }
+    }
+  }
+
   /** Sets the image of the user */
   override def setImage(username: String): ServerServiceCall[Array[Byte], Done] =
     identifiedAuthenticated[Array[Byte], Done](AuthenticationRole.All: _*) {
@@ -403,20 +428,38 @@ class UserServiceImpl(
             }
 
             header.getHeader("Content-Type") match {
-              case Some(contentType) =>
-                if (config.getStringList("uc4.image.supportedTypes").contains(contentType.trim)) {
+              case Some(contentTypeRaw) =>
+                var contentType = contentTypeRaw.trim
+                if (supportedTypes.contains(contentType)) {
                   getUser(username).invokeWithHeaders(header, NotUsed).flatMap {
                     _ =>
-                      database.setImage(username, image, contentType).map {
-                        _ =>
-                          (ResponseHeader(200, MessageProtocol.empty, List(("Location", s"$pathPrefix/users/$username/image"))), Done)
+                      val convertedFuture = if (convertedTypes.contains(contentType)) {
+                        contentType = "image/jpeg"
+                        imageProcessing.convert("jpeg").invoke(ByteString(image))
                       }
+                      else {
+                        Future.successful(ByteString(image))
+                      }
+
+                      convertedFuture.flatMap { converted =>
+                        imageProcessing.fit(profileWidth, profileHeight).invoke(converted).flatMap { profilePicture =>
+                          imageProcessing.thumbnail(thumbnailWidth, thumbnailHeight).invoke(profilePicture).flatMap { thumbnail =>
+                            database.setImage(username, profilePicture.toArray, thumbnail.toArray, contentType).map {
+                              _ =>
+                                (ResponseHeader(200, MessageProtocol.empty, List(("Location", s"$pathPrefix/users/$username/image"))), Done)
+                            }
+                          }
+                        }
+                      }
+                        .recover {
+                          case ex: Throwable => throw UC4Exception.InternalServerError("Error processing the image", ex.getMessage)
+                        }
                   }
                 }
                 else {
                   throw UC4Exception.UnsupportedMediaType
                 }
-              case None => throw new UC4Exception(400, DetailedError(ErrorType.MissingHeader, Seq(SimpleError("Content-Type", "Missing"))))
+              case None => throw new UC4NonCriticalException(400, DetailedError(ErrorType.MissingHeader, Seq(SimpleError("Content-Type", "Missing"))))
             }
         }
     }