+++ product_code/certificate_service/CHANGELOG.md
@@ -1,3 +1,29 @@
+# [WIP v.0.10.4](https://github.com/upb-uc4/University-Credits-4.0/compare/certificate-v0.10.3...certificate-v0.10.4) (2020-XX-XX)
+## Feature
+- The Circuit Breaker ignores now UC4NonCriticalExceptions
+## Refactor
+## Bugfix
+
+# [v.0.10.3](https://github.com/upb-uc4/University-Credits-4.0/compare/certificate-v0.10.2...certificate-v0.10.3) (2020-10-15)
+## Feature
+## Refactor
+## Bugfix
+ - Fixed wrong response code of 202 being sent, instead of 201
+
+# [v.0.10.2](https://github.com/upb-uc4/University-Credits-4.0/compare/certificate-v0.10.1...certificate-v0.10.2) (2020-10-13)
+## Feature
+## Refactor
+## Bugfix
+ - Fix missing OPTIONS call
+ - Add exception handling for hyperledger exceptions
+ - Add better and correct validation for csr
+
+# [v.0.10.1](https://github.com/upb-uc4/University-Credits-4.0/compare/certificate-v0.9.2...certificate-v0.10.1) (2020-10-13)
+## Feature
+## Refactor
+## Bugfix
+ - Fix missing enrollment
+
 # [v.0.9.2](https://github.com/upb-uc4/University-Credits-4.0/compare/certificate-v0.9.1...certificate-v0.9.2) (2020-10-09)
 ## Feature
 ## Refactor